+++ tests/phpunit/bootstrap.php
@@ -25,6 +25,16 @@ if ( file_exists( $polyfill_path ) ) {
 ini_set( 'display_errors', 'on' );
 error_reporting( E_ALL );
 
+// Backward compatibility (PHPUnit < 6).
+$phpunit_backcompat = array(
+	'\PHPUnit\Framework\TestCase' => 'PHPUnit_Framework_TestCase',
+);
+foreach ( $phpunit_backcompat as $new => $old ) {
+	if ( ! class_exists( $new ) && class_exists( $old ) ) {
+		class_alias( $old, $new );
+	}
+}
+
 $_tests_dir = getenv( 'WP_TESTS_DIR' );
 
 // Check if installed in a src checkout.